[P1.3] haven-git: treebuilder insert expects i32 not ObjectType enum

+++ crates/haven-git/src/lib.rs
@@ -262,7 +262,7 @@ impl VaultWriter {
         index.clear()?;
         let blob_oid = repo.blob(&write.new_content)?;
         let mut builder = repo.treebuilder(None)?;
-        builder.insert(Path::new(&relative), blob_oid, ObjectType::Blob)?;
+        builder.insert(Path::new(&relative), blob_oid, ObjectType::Blob as i32)?;
         let tree_oid = builder.write()?;
         let tree = repo.find_tree(tree_oid)?;
         write_commit(&mut repo, &tree, &write, &self.inner.identity)
